+++ source/standalone_examples/api/isaacsim.sensors.camera/camera_add_depth_sensor.py
@@ -17,18 +17,17 @@ from isaacsim import SimulationApp
 
 simulation_app = SimulationApp()
 
-from isaacsim.core.utils.stage import get_current_stage
+from isaacsim.core.experimental.utils.stage import define_prim, get_current_stage
 from isaacsim.sensors.camera import Camera, SingleViewDepthSensorAsset
 
 # Create a root XForm on the stage
-stage = get_current_stage()
-root = stage.DefinePrim("/root", "Xform")
+root = define_prim("/root", "Xform")
 
 # Create a new Camera prim on the stage
 camera = Camera(prim_path="/root/Camera")
 
 # Create a new Scope to contain the template render product for the depth sensor
-scope = stage.DefinePrim("/root/TemplateRenderProduct", "Scope")
+scope = define_prim("/root/TemplateRenderProduct", "Scope")
 
 # Create a new template render product, applying the OmniSensorDepthSensorSingleViewAPI schema and setting custom attributes
 depth_sensor_attributes = {
@@ -40,7 +39,8 @@ render_product = SingleViewDepthSensorAsset.add_template_render_product(
     **depth_sensor_attributes,
 )
 
-# Set default prim as the camera
+# Set default prim
+stage = get_current_stage()
 stage.SetDefaultPrim(root)
 
 # Export the stage