+++ source/extensions/isaacsim.replicator.behavior/docs/CHANGELOG.md
@@ -1,4 +1,7 @@
 # Changelog
+## [1.3.4] - 2026-01-21
+### Changed
+- Increased test tolerances for golden image comparison until randomizations are deterministic 
 
 ## [1.3.3] - 2026-01-22
 ### Changed